chore: moved all the types
