+++ src/app/configs/config.ts
@@ -1,5 +1,5 @@
 import common from 'app/configs/environments/common';
-import type { EnviornmentConfig, Config } from 'app/configs/types';
+import type { EnviornmentConfig, Config } from 'types';
 const environment: EnviornmentConfig = require(`app/configs/environments/${process.env.NODE_ENV}`);
 
 export const config: Config = {
