+++ src/app/configs/environments/development.ts
@@ -1,4 +1,4 @@
-import { EnviornmentConfig, HealthChecksConfig, KnexConfig, MysqlConnectionString } from 'app/configs/types';
+import { EnviornmentConfig, HealthChecksConfig, KnexConfig, MysqlConnectionString } from 'types';
 
 import Knex from 'knex';
 import os from 'os';
